tracksupport/pyproject.toml
[project]
name = "tracksupport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "av>=16.1.0",
    "debugpy>=1.8.20",
    "ffmpeg>=1.4",
    "hydra-core>=1.3.2",
    "imageio>=2.37.2",
    "lap>=0.5.13",
    "omegaconf>=2.3.0",
    "opencv-python-headless>=4.13.0.92",
    "peft>=0.18.1",
    "positional-encodings[pytorch]>=6.0.3",
    "pytorch-ignite>=0.5.3",
    "scikit-learn>=1.8.0",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "torchcodec>=0.10.0",
    "transformers>=5.2.0",
    "ultralytics>=8.4.18",
    "wandb<0.18.0",
]

[tool.uv]
override-dependencies = [
    "opencv-python; python_version < '0'",
    "opencv-contrib-python; python_version < '0'"]
